Refactor ScmConfig
