+++ src/main/java/nl/tudelft/jenkins/jobs/GitScmConfig.java
@@ -12,17 +12,18 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * Date: 1/21/14
  * Time: 4:17 PM
  */
-public class GitScmConfig implements ScmConfig {
+public class GitScmConfig extends ScmConfig {
     private static final String XPATH_SCM_GIT_URL = "//scm/userRemoteConfigs/hudson.plugins.git.UserRemoteConfig/url";
     private final static String GITCONFIG_XML_PATH = "/nl/tudelft/jenkins/jobs/GitScmConfig.xml";
 
-    private String address;
 
     public GitScmConfig(String address) {
-        checkNotNull(address, "address must be non-null");
-        this.address = address;
+        super(address, null, null);
     }
 
+    private GitScmConfig(String address, String username, String password) {
+        super(address, username, password);
+    }
 
     @Override
     public Element getXmlContent() {
@@ -32,29 +33,4 @@ public class GitScmConfig implements ScmConfig {
         return doc.getRootElement().detach();
     }
 
-    @Override
-    public String getUsername() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    @Override
-    public String toString() {
-        return "GitScmConfig{" +
-                "address='" + address + '\'' +
-                '}';
-    }
 }
